Create utills classes to seperate out logic for different schedule types
Summary:
Create a helper class that provides functions that differ for different schedules
and provide this class as a member of the pipeline visitor.

Chose to try and contain this all in the utills section instead of as virtual functions
of the pipline visitors but could easily change. Also in the utils section choose
to use a variant instead of virtual functions as meant could keep the templating
the same a before, otherwise would have to template the class.

Also to keep all the declarations in the utils header instead of the pipeline
visitor header used a unique ptr and forward declared it even though don't need this
as not using virtual functions. Downside is an extra heap allocation.

Reviewers: jakeh, #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: jakeh, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: markf, georgep

Maniphest Tasks: T34310

Differential Revision: https://phabricator.sourcevertex.net/D46292