Use poplibs outputFullSequence instead of doing a slice for all RNNs
Summary:
Give the LSTM and basic GRU ops an output_full_sequence attribute which is used to populate the poplibs outputFullSequence option.
The dynamic GRU and AUGRU aren't given the attribute since outputFullSequence=false isn't currently supported with them in Poplibs.

When setting this flag to false, all the outputs are returned in the intermediates instead. If the output shapes of a GRU with this option set to true are [s, bs, hidden] (outputs) and [s, 3, bs, hidden] (intermediates) then the output shapes become [bs, hidden] and [s, 4, bs, hidden] when it's set to false. Most of the changes in this diff are to accommodate these shape changes.

The Keras layers exposed this option to the user by doing a slice on the output tensor. Now the Keras layers pass the option down to C++ and don't do a slice. The rest set it to True which is the poplibs default.

This allows us to take advantage of any current or future poplibs optimisations which exploit the outputFullSequence flag.

TF2.1 Only
Ref T37173

Test Plan:
CI - there should be no change in behaviour for the tests that toggle the option.
Also added a couple tests in the existing gru_test and lstm_test files.

Reviewers: gauthamg, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T37173

Differential Revision: https://phabricator.sourcevertex.net/D45900