Add in-memory executable cache
Summary:
Cache for compiled Poplar core executables.

Each key is a function which identifies a unique XLA computation and the IPU
system configuration it was compiled with and the device ordinal it is being
compiled for, and the value is executable generated by compiling the
computation.

The cache is implemented as a singleton which means that the core executables
can be shared for executables being compiled in different sessions.

When an executable is generated by GetOrCompileExecutable, the
PoplarExecutableCore part of the executable (the part which can be shared) is
stored in the cache and the compiled PoplarExecutable has a weak reference to
a PoplarExecutableCore.

The choice of the key means that PoplarExecutableCore can only be shared when
compiling for the same device ordinal - this is required because the compiled
Poplar Engine is not thread safe and it cannot be shared between multiple
runtimes executing concurrently.

This class is threadsafe - if there are parallel compilation requests for the
same hash key, only a single PoplarExecutableCore will be compiled whilst
other threads wait.

Note that a PoplarExecutable stores a shared ptr to PoplarExecutableCore,
once all of these are destroyed, the weak reference in the cache will become
expired. If the executable is required again then it will be recompiled from
scratch.

Fix T40217

TF2.1 Only

Test Plan: CI, clear_xla_cache_test passes

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk, hakons

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, babakk

Maniphest Tasks: T40217

Differential Revision: https://phabricator.sourcevertex.net/D46790