Move AUGRU test into gru_test
Summary:
Previously, the AUGRU test was not enabled (it didn't have the googletest.main() call and it wasn't in the BUILD file test list) and the CPU comparison numbers were actually running on the IPU.

This diff moves the AUGRU test into the gru_test file, where the normal and dynamic GRUs are tested. This means any parameterization we do of gru_test in future can benefit the AUGRU test coverage too.

TF1.15 Only (The test is in the test suite in this branch but not in the 2.1 branch)
Ref T23681

Test Plan: New test passes.

Reviewers: gauthamg, yanw, #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T23681

Differential Revision: https://phabricator.sourcevertex.net/D46847