Added unit test for nested outputs
Summary:
Currently, when using on-chip synthetic data and Keras, with nested lists of output
tensors, an error is raised where we try to get the shape of the tensor. This is because
the content of the variable is the inner list - not an actual tensor.

This diff flattens the nested list before iterating over it, to ensure that each element
is a tensor.

TF2.1 Only

Test Plan:
Added a test for this specific case to tensorflow/python/ipu/tests/keras/keras_synthetic_data_test.py.

Tested without fix - error appears. With fix error is gone.

Reviewers: georgep, #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: georgep, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T41278

Differential Revision: https://phabricator.sourcevertex.net/D47375