Improve auto scheduler selection and include SHORTEST_PATH
Summary:
Run the HLO alias analysis and pass the result to the heap simulator.
With the empty alias analysis that we used before the heap simulator
always returned 0 and we chose the first scheduler (POST_ORDER),
while now we let it choose among the three POST_ORDER,
CLUSTERING and SHORTEST_PATH.

LOOK_AHEAD is still not added in auto mode because of its time
complexity (it took 4 minutes with default settings on BERT-Base).

The scheduler is hard coded in some tests that relied on a specific schedule.

Ref. T40113
Ref. T37032

TF2.1 Only.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh, georgep, markf

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, markf

Maniphest Tasks: T40113

Differential Revision: https://phabricator.sourcevertex.net/D48010