Initial structure + tests for replica identical data flow analysis.
Summary:
Adding initial infrastructure for a visitor based dataflow analysis to identify replica indentical values in a computation. The only tuple shaped instructions that are explicitly supported atm are tuple and get-tuple-element. I'll follow up with a separate diff to extend it with the other ones we need.

As part of this I moved what was EarlyFuseTest and its utilities into test_utils.h so it can be reused in other tests

Test Plan: New C++ tests

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh, samuelh

Maniphest Tasks: T41163

Differential Revision: https://phabricator.sourcevertex.net/D47901