Add an argument to mask outfeed outputs in pipelining
Summary:
Currently the last computational stage will always output all the returned values
to the outfeed queue. In the new keras implementation we want to output some values
to the optimizer function, but not to the actual outfeed so we can set this mask.

Ref T41124

Test Plan: CI, added new tests

Reviewers: jackh, jakeh, #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: jackh, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T41124

Differential Revision: https://phabricator.sourcevertex.net/D48400