Adding support for simple calls to replica dataflow analysis
Summary:
This change adds support for `call` instructions with the Function CallConfig to the replica dataflow analysis. When encountering one of these we perform the dataflow analysis on the HloComputation its calling, but with the operands of the call instead of the actual parameters of the computation. This requires that that the HloModule have a flattened call graph, so after visiting the sub computation we can just move the results into the calling visitor without conflict. We then use the analysis results for the sub computation root as the result for the caller.

TF1 Diff - D48640

TF2.4 Only

Test Plan: C++ tests

Reviewers: #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh, jakeh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh

Subscribers: markk

Maniphest Tasks: T41162

Differential Revision: https://phabricator.sourcevertex.net/D48358