Update tensorflow to use poplar engine names for profiling report directory
Summary:
Tensorflow now passes a name to the engine/executable constructor which will be used to create a subdirectory of autoReport.directory to put the profile.pop into.

This then replaces the code in tensorflow that set the autoReport.directory and the user can now set it again.

The handling of the frameworks.json is a bit message, as the exisiting logic put that in the module report directory.

I would like to move that up to the autoReport.directory now and for it too contain information for the application

Test Plan: All exisiting unit tests pass./

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jamiep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jamiep

Subscribers: jamiep, george, georgep

Maniphest Tasks: T42277

Differential Revision: https://phabricator.sourcevertex.net/D48151