Add gcl::CollectiveBalancedHostRearrangement support
Summary:
This commit is one of the series in attempt to split larger CBR change into more controllable/reviewable smaller changes.

This commit provides serialisation and host rearrangement support in poplar executor.
Ref T35351.

Test Plan:
CI

This code shouldn't be used by any existing code, so we test only that it's not breaking anything.
In order to add end-to-end testing, on-device rearrangement will be required, and it will be added in CBR change.

Reviewers: #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh, georgep, hakons

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh, georgep, hakons

Maniphest Tasks: T35351

Differential Revision: https://phabricator.sourcevertex.net/D48777