Inferring allocations from other allocating instructions
Summary:
The current allocation finder behaviour is that every allocating instruction will fully traverse the subtree of it’s consumers to build up a target.

These changes make it so if you encounter another allocating instruction while traversing the subtree of consumers, rather than traversing through it, we infer what we would find in that branch of the tree from that instructions target.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh, hakons

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh

Subscribers: jackh

Maniphest Tasks: T38658

Differential Revision: https://phabricator.sourcevertex.net/D46809