Adding StochasticRoundingBehaviour enum to Config proto.
Summary:
Changing type of FloatingPointBehaviour.esr member, which controls stochastic rounding, from bool to an enum so we can support the existing true/false behaviour and our new 'only for replica identical instructions' behaviour.

Python tests don't need to be updated for this change yet as the semantics of setting the esr member with a bool hasn't change - the bool is implicitly converted to an int.

Test Plan: CI

Reviewers: #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markf, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T41158

Differential Revision: https://phabricator.sourcevertex.net/D48958