Add application compile op
Summary:
This adds the `IPUApplicationCompile` op that compiles
a function and returns a string to the compiled executable.

It also supports freezing the variables such that they become
constants embedded into the executable.

This is not really tied to the application runtime per se, but it seems
good to limit the initial scope.

Ref. T41635.

TF2.4 only.

Test Plan: Added new tests and integrated with existing application runtime test.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, gauthamg, jakeh, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh, georgep

Subscribers: georgep

Maniphest Tasks: T41635

Differential Revision: https://phabricator.sourcevertex.net/D48034