Updating replica dataflow analysis to support visting conditionals/loops.
Summary:
Updating dataflow analysis to support conditionals/loops.

When dealing with conditionals we say that the instructions value is identical across replicas if the predicate is and the corresponding value from each branch always is, since we can't tell ahead of time which branch will be taken.

For repeat loops we say that a value is identical if it's identical in both iter0 and iter1 (when it has more than 1 iteration), as the loop parameters change from being the initial caller operands to the result of the previous iteration. If it only has 1 iteration then we treat it as a function call.

For while loops we say that a value is identical if it's identical in the initial caller operands, and both iter0 and iter1 of the body, since we don't know how many iterations there will be. For the same reason we also require that the conditional be identical when called with its initial parametrs and when called with the results of iter0 and iter1, so we know whether the replicas perform the same number of iterations.

Subcomputations for conditionals/loops are only visited if their predicate/conditional are replica identical., otherwise we set all the instructions in the subcomputation to be differing. This check is needed since different replicas could execute the subcomputations a different number of times (or not at all).

TF1 Diff - D48923

TF2.4 Only

Test Plan: Extended C++ HLO tests

Reviewers: #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh, jakeh, vladimirm

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh, vladimirm

Subscribers: markk

Maniphest Tasks: T41162

Differential Revision: https://phabricator.sourcevertex.net/D48672