Adding support for pipelines and tuple select statements to replica dataflow analysis
Summary:
This change adds support for Pipeline/PipleineStages and TupleSelect instructions.

A Pipeline call is iterative and handled in the same way as a repeat loop, so depending on its repeatCount it's visited either once (with the pipelines initial arguments) or twice (with the initial arguments and then with the categories that generates).
A PipelineStage call is treated like a normal function and just visited with its callers arguments.

A tuple-select instruction is handled like a conditional - the values are only replica identical if the select predicate is and if they're in both tuples.

The DefaultAction implemention correctly handles regular select instructions.

Test Plan: C++ HLO tests

Reviewers: #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh, samuelh, vladimirm

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh

Maniphest Tasks: T41162

Differential Revision: https://phabricator.sourcevertex.net/D49080