Adding ReduceMany op and colocation helper
Summary:
TF1.15 Only

Adding hlo-only ReduceMany op.
Reductions (including fusion reductions) can be combined into these ReduceMany ops.
The interface to control this is this is a new ipu config option optimizations.maximum_reduce_many_buffer_size.
This also has a change to the clustering scheduler to prevent a memory regression in a test. The change makes it so that ops with a valid colocator helper will not be put into their own cluster if the buffer size for the colocator is zero. This will prevent colocator helpers added in future from affecting the schedule in unrelated models.

V2 Diff: D48529

Test Plan:
Tests check that that simple reduces and reduce fusions get combined in hlo according to the specified optimizations.maximum_reduce_many_buffer_size.
There is also a test which executes a graph with a ReduceMany and checks the output values.

Reviewers: #tensorflow, simonl, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: zhenyingl, mariok, georgep

Maniphest Tasks: T40084

Differential Revision: https://phabricator.sourcevertex.net/D49202