Only register application compile op for CPU
Summary:
The trick we tried to pull by pretending to support the compile op on
IPU backfired when freezing the variables, as they were not anymore
host-memory tensors as required by the XLA compiler (ref.
docstring for `XlaCompiler::constant_value`).

This fixes a crash that could happen when freezing the variables
when the compile op was placed on the IPU, and instead gives
an error message telling that it must be placed on the CPU.

Fixes T44934.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, markf

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, markf

Maniphest Tasks: T41633, T44934

Differential Revision: https://phabricator.sourcevertex.net/D50304