Dummy data feed refactor.
Summary:
Use the timeout to limit how long we wait on the `PeekInputData` condition variable, when it's possible to deadlock.

What's changed:
- The poplar callback now waits for the timeout, then processes dummy data.
  - Removed dummy thread.
- Added some more tracing to make it clearer when we are blocked.

The consequence of this is we will never have more dummy items in the queue than the device explicitly requests.

This is what a timeout looks like. Waits 500us (the chosen timeout) then pushes a dummy batch, the subsequent callbacks don't wait.
{F1652440}

Resolves T44892

Test Plan: Existing HW CI tests

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, davidn

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, davidn

Maniphest Tasks: T44892

Differential Revision: https://phabricator.sourcevertex.net/D50552