Improve error message for unsupported RTS config
Summary:
Previously, when trying to do replicated partitioning with multiple instances per
IPU-link domain, it would fail in GCL with an error like:
```
[Poplar][Compile engine] poplar_error: Can only send to corresponding replica over GW links
```
This new error will hopefully give the user a better idea about what is going on and what
can be done to resolve the issue until GCL adds support for this.

Test Plan:
Added new test. Tested manually in on a multi-ILD system with poprun.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T35352

Differential Revision: https://phabricator.sourcevertex.net/D51192