Adding StochasticRoundingMethod backend option
Summary:
This change adds the `StochasticRoundingMethod` option to the backend config. This is intended to be used as an explicit way of describing how we want to perform stochastic rounding (with an identical seed/differing seed or either). By having an extra backend option we avoid having to overload the meaning of being replica identical and having to add an extra category to the replica dataflow analysis, which will further complicate it.

`StochasticRoundingMethod` gets set by the `AddStochasticRoundingOptions` so only instructions which require a specific type of seed will cause the seeds to be changed. It's currently setup so that instructions which read/restructure data don't change the seed.

(TF1 - D51636)

TF2.4 Only

Test Plan: C++ Tests

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jakeh, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T45895

Differential Revision: https://phabricator.sourcevertex.net/D51361