Allow dynamic iterations for non grouped pipeline visitors
Summary:
Before this took a compile time int, now takes a variant of tensor or int.
In the int case calls same as before. For the tensor wraps each sequence
in an If program (Note I think this could be more efficient with a big switch
 if poplar outlining is good, which I think it is).

For recomputation sequences for the cases it wokrs for the number of
iterations is always zero. Otherwise hits error inside MaskOutEmpty.

Test Plan: modified pipeline vistor tests. Python tests in different diff

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh

Maniphest Tasks: T41148, T41149, T41150

Differential Revision: https://phabricator.sourcevertex.net/D51030