Fixing prng seed inconsistencies with repeat loops.
Summary:
This change updates RepeatLoopVisitor to make additional calls to ChangeStochasticRoundingMethod to try and make sure that the seed values remain consistent while running the loop. Inconsistency can happen when at runtime the seed type is different to what was set while lowering the hlo. This can happen when the order of operations in the generated poplar program differs to the order in which the corresponding hlo instructions were visited. In the case of repeats this happens within the loop, since the loop is only visited once but executed multiple times, and when performing a resource update, since this is explicitly reordered.

As part of this change I added a flag to `CompilerResources` to control whether not we perform consistency checks and updated PoplarXlaFlags to support being reloaded, both of these are for testing purposes only.

Test Plan: New C++ tests.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jakeh, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Maniphest Tasks: T47570

Differential Revision: https://phabricator.sourcevertex.net/D53325