Discard constant placeholders when compiling
Summary:
Discard constant placeholders created by earlier execution of constants in
eager mode when we compile a function for graph mode.

We already have these constants captured by value, so we can safely
remove these from external captures.

Fixes an issue where `InvalidCaptureException` is raised when a constant
is used as an operand to a TF op in a `PipelineStage`.

TF2.4 Only

Test Plan: Added a test based on the reproducer given in the bug report.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jackh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jackh

Subscribers: jackh, georgep

Maniphest Tasks: T45648

Differential Revision: https://phabricator.sourcevertex.net/D53647