Adding test for using PrngSeedState with a sharded/replicated graph.
Summary:
This change adds a test for changing the prng seeds in a task parallel graph, it's intended to very simply emulate the kind of execution we might expect in TF when using pipelines.

Even though the seed changes are done on the master graph it doesn't look like this introduces any additional syncs, although it adds some redundant state changes but we should be able to remove those after everything is working.

Test Plan: New tests.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, georgep, jakeh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Maniphest Tasks: T47570

Differential Revision: https://phabricator.sourcevertex.net/D53795