Remove ipu_compile_summary and everything that uses it
Summary:
When we switched to the IPUConfig, turning on IPU events was made internal by putting it on a hidden _profiling category of the config.
Anyone who wants to use the ipu_compile_summary op (or anything that uses it) would have to know that:
1. events need to be enabled in the config for it to produce anything
2. the IPUConfig._profiling category exists
neither of which are documented, and it hasn't been requested. Hence the ipu_compile_summary was implicitly, transitively deprecated when the IPUConfig was introduced.
There is only one usage of it in the examples which was broken and is being removed in D53515.
We also don't use this op internally in any tests (which aren't testing the summary itself), so we can remove it instead of making it internal as it's unused.

Ref T47845
Depends on D53515 (no external users of the op)
Depends on D53807 (to enable removal of get_ipu_reports)

TF2.4 Only.
TF1 diff - D53516

Test Plan: CI. Removed two tests that test the summary itself.

Reviewers: alfiee, georgep, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: georgep, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T47845

Differential Revision: https://phabricator.sourcevertex.net/D53618