Use variable handle rather than name when exporting variables to PopEF
Summary:
TF v2 doesn't pass variable names to the C++ side, instead using anonymous variable handles
to refer to the variables (the exception to this is when running with v2 features disabled).
This patch changes the variable export to use handles, so that the names in the variable
export match the anchors in the executable metadata export.

The disadvantage of this is that the variables will now be stored in the PopEF file with
names like "_AnonymousVar0", rather than the names they have been given by the user.
There's no way that I can find to fix this without adding a lookup table to TF to enable
resource handles to be resolved back to the original variable names. If we want to do this,
this patch should be rejected and I'll do another patch with the lookup table.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, anthonyb, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T45630

Differential Revision: https://phabricator.sourcevertex.net/D53392