Fix prng inconsistencies with pipelines
Summary: Tweaking prng seed setup during pipeline lowering so that after each stage/recomputation we go back to a specific seed. This is intended to keep the values consistent as the pipeline gets executed since the programs in a pipeline can be executed in a variety of different ways (depending on the schedule). We don't do anything for the other program types (infeeds/copies/outfeeds) as we don't expect those to require any seed changes.

Test Plan: CI + New test

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, samuelh, jakeh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Maniphest Tasks: T47570

Differential Revision: https://phabricator.sourcevertex.net/D54052