Move IPUConfig _profiling.enable_ipu_events to a test utility
Summary:
We want to remove IPUConfig._profiling. The only field that we still depend on is enable_ipu_events. It's now serving as an internal-only testing tool thanks to previous deprecation/efforts, so I've moved it to test_utils and turned it into a callable function. This way, the existing user-facing APIs don't change and we can remove _profiling in a future diff.

The callable function works by wrapping the config's _create_protobuf, so that it enables the IPU events on the protobuf it creates.
To wrap _create_protobuf, we have to reassign the instance's method to the wrapped version. We have to use `object.__setattr__` rather than `instance.__setattr__`, as IPUConfig overrides `__setattr__`. Normally, re-assigning an instance's method would mean rebinding the new method to the instance, but we don't need to access the instance inside the wrapped method so it's fine.

Ref T39600
Depends on D53516 (remove user-facing usage of events)
Depends on D53570 (fix branch divergence in batch norm test)
Depends on D53571 (fix branch divergence in batch norm test)

TF1.15 Only. (2.4 has some Keras tests that need changing).

Test Plan: Refactor so CI should pass.

Reviewers: alfiee, georgew, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Maniphest Tasks: T39600

Differential Revision: https://phabricator.sourcevertex.net/D53414