Add `to_{dict,json}` and `__repr__` to IPUConfig
Summary:
This commit adds the `to_dict()` and `to_json()` methods to the
IPUConfig class, allowing serialising the configuration object's
current configuration to another format, as well as corresponding
`from_dict()` and `from_json()` methods which allow complete or partial
restoration of configuration from a flattened dict or JSON object.

Additionally, implementing `__repr__()` now mimicks the existing
behaviour with the `IpuOptions` object when `print()`-ing or evaluating
the object in the Python REPL to display a simple summary of all the
configuration retained within the object.

As `_create_protobuf()` is only useful on `IPUConfig`, it has been moved from
`_ConfigBase` to `IPUConfig`.

Test Plan:
Added several tests which ensure a subclass of _ConfigBase can be both
serialised and deserialised to a dict or a JSON string depending on a
`is_json` parameter. Two special tests also check that enum values can
be serialised as ints, and then rehydrated on import.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, georgep, markf

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, markf

Maniphest Tasks: T43332

Differential Revision: https://phabricator.sourcevertex.net/D52172