Updating ReplicaIdenticalOnly mode to use StochasticRoundingMethod_None.
Summary: Originally the intention was for ReplicaIdenticalOnly mode to set the stochastic rounding backend option to off for those instructions which are not replica identical, however since we only use that option from BaseVisitor::Preprocess we'd have to update loops/conditional/pipeline lowering to make sure that the tracked stochastic rounding state was consistent with the runtime one. By using StochasticRoundingMethod_None we avoid that, as that work has already been done for the other StochasticRoundingMethods. The downside is that we now have two ways of setting SR, until the experimental prng stability flag is made the default.

Test Plan: CI

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep, gauthamg

Maniphest Tasks: T48700

Differential Revision: https://phabricator.sourcevertex.net/D54541