Implement GradientAccumulatorAddWithScale
Summary:
Replace GradientAccumulatorAdd with GradientAccumulatorAddWithScale, which takes in an additional scale parameter for scaling the accumulator value prior to accumulation. The objective is to enable accumulation of the type `acc <- acc * acc_scale + grad * grad_scale`, which will enable us to support a running mean. In this implementation, `grad * grad_scale` will be computed in Python and passed to the new op along with `acc_scale`

REF T46005
TF2.4 Only
TF1 diff: D53583

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, vladimirm, samuelh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, vladimirm, samuelh

Maniphest Tasks: T46005

Differential Revision: https://phabricator.sourcevertex.net/D55699