Use OpDef rather than Tin to determine original tensor types when rearranging resources for XLA
Summary:
tf2xla rearranges resources to the end of the input/output lists as it's a requirement for XLA.
To do this, the DataType of all the input/output tensors needs to be determined.
For upstream, there is only one .Input/.Output on the OpDef(Builder) for While and If, which uses a Tin/Tout attr to set the types. They then easily use this attribute to determine the types.

For multiple .Inputs/.Outputs, this shortcut can't be used. This diff formalises how we find the types of the input/output tensors by iterating through all the ArgDefs on the OpDef and building a list of DataTypes, using the upstream ArgNumType utility function.

Ref T48256

Depends on D56891 (Rearrangement wasn't being done because of incomplete type finding. With this diff, rearrangement is done, but GA count is duplicated so it fails).

Test Plan: Tested locally with a version of the PipelineOp with two .Inputs and a situation where resources are given as non-terminal inputs. Gives a segfault without the diff and passes with the diff. I would add this as a test but I'm going to change the PipelineOp to have one .Input for unrelated reasons.

Reviewers: babakk, georgew, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: georgew, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep, jackh

Maniphest Tasks: T48256

Differential Revision: https://phabricator.sourcevertex.net/D56886