Absorb GA count Input into first Input in the Pipeline OpDef
Summary:
Currently, the GA count is passed as an input to the Pipeline twice: once through the captured_args and another through the gradient_accumulation_count Input in the OpDef.

If upstream ever needs to convert the Graph that this creates into a FunctionDef via GraphToFunctionDef (for example, during resource rearrangement when the function is rebuilt with the permuted inputs), we find that GraphToFunctionDef cannot handle _Args being used multiple times as input to the same Node. This is because that function only iterates over the nodes once to populate the function arguments.

To deduplicate the GA count input, this diff removes it as an explicit input in the OpDef, and trusts the user to supply it as the first input in the input list (as it already was).
We lose the type checking from the OpDef, but add it back into the OpKernel ctor.

Ref T48256

Test Plan: Tested locally with a Pipeline with a non-terminal resource (which therefore triggers resource rearrangement and calls GraphToFunctionDef). Fails without the diff, passes with.

Reviewers: samuelh, georgep, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: georgep, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T48256

Differential Revision: https://phabricator.sourcevertex.net/D56891