Support `@classmethod` in KerasExtension
Summary:
This change adds the ability to override a method annotated with the
`@classmethod` decorator within a `KerasExtension`.

TF2.4 Only

Test Plan:
Added a new test in `extensions_test.py` to ensure the KerasExtension
override is run for a given classmethod also annotated with
`base_layer.extension_delegate` when inside an `IPUStrategy`.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jackh

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep, jackh

Subscribers: jackh, georgep

Maniphest Tasks: T50790

Differential Revision: https://phabricator.sourcevertex.net/D57246