Removing redundant masks from triangular shape ops
Summary: Adds a pass which removes masking from cholesky and triangular-solve. In the TensorFlow implementation of these ops, the zero part of the result can contain junk values which often need masking. This is not the case with the poplibs implementations of these ops, so we can safely remove masking. Masking is added internally by tf.linalg.cholesky so the pass is required to remove masking in that case.

Test Plan:
Added tests for 4 different arrangements of masking (two cholesky and two triangular solve).
All 4 valid comparison directions are covered, as well as lower=true and lower=false, and both variations of parameter order in select.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, georgep

Subscribers: georgep

Maniphest Tasks: T51283

Differential Revision: https://phabricator.sourcevertex.net/D57446