Adding support for setting BufferUseKind on HloPoplarBuffer objects.
Summary: Adding support for setting BufferUseKind on HloPoplarBuffer objects. This is managed by the HloPoplarBufferSet that contains it, so that as a buffer sets use kind changes then so do the contained buffers. Like a buffer set though the highest usage kind is preserved.

Test Plan: CI + New tests

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, georgep

Maniphest Tasks: T52583

Differential Revision: https://phabricator.sourcevertex.net/D57913