Check for missing plans (slice ops) in EmbeddingPlansPreplanning
Summary:
Count the slice ops in both the flattened module and the incoming pass HLOModule.
Issue a "LOG(FATAL)" if these do not match.
This change does NOT address the fundamental issue/gap as described in the parent task; it only improves behaviour somewhat by trapping the error earlier (at the point of failure).

Test Plan:
Ran the reproducers provided in the parent task and confirmed that this change will abort immediately if there is a mismatch (disagreement in slice op counts).
In addition, with TF_CPP_MIN_VLOG_LEVEL=1, slice ops that are missing in the flattened module will be printed.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, georgep

Subscribers: georgep

Maniphest Tasks: T52784

Differential Revision: https://phabricator.sourcevertex.net/D58064