Add documentation for infeed/outfeed shorthand syntax
Summary:
This diff adds information to the TF2 docs about the changes to in/outfeeds in SDK 2.2+.

It explains that the user can simply wrap the dataset in an iterator and access that from within the training loop.
It also adds a section which describes how to iterate over an outfeed, with a short example taken from George's
Notebook examples for TF 2.4.

TF2.4 Only

Reviewers: #tech_docs, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamiep, georgep, markf

Reviewed By: #tech_docs, #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamiep, markf

Subscribers: markf, tims, georgep, jamiep

Maniphest Tasks: T46827

Differential Revision: https://phabricator.sourcevertex.net/D57992